Fix conda build 20230809 (#24)
* Update README with conda-forge package notes

* Fix translate_tabular.cpp; spdlog template issues with atomic<int>

* Update README for conda-forge

* Fix cmake to properly use USER_INSTALL_DIR for linux

* Reorganize conda-forge sectino of README

+++ cpp/translate_tabular/src/translate_tabular.cpp
@@ -81,7 +81,7 @@ int TranslateTabular::Translate()
             status_string = "SUCCESS";
             SPDLOG_INFO("TranslationManager associated with thread {:d} "
                 "status: {:s}, exit code {:d}", (*it)->thread_index, 
-                status_string, (*it)->exit_code);
+                status_string, static_cast<int>((*it)->exit_code));
         }
         else
         {
@@ -91,7 +91,7 @@ int TranslateTabular::Translate()
 
             SPDLOG_ERROR("TranslationhManager associated with thread {:d} "
                 "status: {:s}, exit code {:d}", (*it)->thread_index, 
-                status_string, (*it)->exit_code);
+                status_string, static_cast<int>((*it)->exit_code));
         }
     }
     if (exit_code != 0)
